fix(autoware_behavior_velocity_planner): fix bugprone-narrowing-conversions warnings (#939)

+++ planning/behavior_velocity_planner/autoware_behavior_velocity_planner/src/node.cpp
@@ -404,7 +404,7 @@ void BehaviorVelocityPlannerNode::publishDebugMarker(const autoware_planning_msg
   for (size_t i = 0; i < path.points.size(); ++i) {
     visualization_msgs::msg::Marker marker;
     marker.header = path.header;
-    marker.id = i;
+    marker.id = static_cast<int>(i);
     marker.type = visualization_msgs::msg::Marker::ARROW;
     marker.pose = path.points.at(i).pose;
     marker.scale.y = marker.scale.z = 0.05;
